Found another double occurance I needed to change

+++ src/cli_evaluator/main.cpp
@@ -31,7 +31,7 @@ int main(int argc, char* argv[]) {
     Graph inputGraph = GraphIO::readEdgeList(options.edgeListPath, options.edgeListComment, options.edgeListDelimiter);
 
     // read in embedding
-    std::vector<std::vector<double>> coords = EmbeddingIO::readCoordinatesFromFile(
+    std::vector<std::vector<float>> coords = EmbeddingIO::readCoordinatesFromFile(
         options.embeddingPath, options.embeddingComment, options.embeddingDelimiter);
     std::shared_ptr<Embedding> embedding = EmbeddingIO::parseEmbedding(options.embType, coords, options.lPNorm);
     if (embedding == nullptr) {
